Add files via upload
Provided a ReadMe to act as user instructions.  Created a fix for the database access by including an address input bar within the app for the user to include the location of their version of the TextTile.mdf file.  It should be noted that this is a simulation with mock controls, jobs, and database, but it should be functional with some minor edits and access to the appropriate materials and specs.